refactoring to prepare rendering of different layers for glyphs
The centerline and outline are now rendered in different layers.
The centerline is styled as a green line in the less/css.
The future may introduce switching different layers on/off and
layers that provide more information about the glyphs (points,
cps-drilldown, controls, stuff like that)

Use setTimeout instead of angulars $interval