Add a Selector constructor operator to the formulae enginge.
Detect when a "final" Selector does not select anythin in CPS.

Whitelist the "id" MOM Node property for cps (which is useful to dynamically
create Selectors with the Select operator)

Using the glyph.id in the cps generators for metapolation and interpolation
relaxes the glyph order requirement for interpolation compatibility. We no
mereley need glyphs to have the same name, instead of the same position
within their masters.