Interpolate angles via vectors from polar cordinates and extract the final angle.
The previous approach with normalizeAngle was not working very well.
E.g.: interpolating two angles one 0 rad and one 2 * PI - 0.000001 rad
would result with 50/50 interpolation at ~ PI, which is almost exactly
the opposite direction

NOTE: This still has an edge case two where it fails! When The
interpolated Vectors end up having no magnitude, the angle
returned is just 0 which is probably as wrong as any other number
in that case.